requirements.txt
streamlit
streamlit_option_menu
pyserial